+++ src/navigation/PageLayout.jsx
@@ -1,13 +1,23 @@
 import React from 'react';
 import NavigationBar from "./NavigationBar";
 import {Outlet} from "react-router-dom";
+import {useSelector} from "react-redux";
+import {getTheme} from "../components/themes/customTheme";
+import {CssBaseline} from "@mui/material";
+import {ThemeProvider} from "@mui/material/styles";
 
 export default function PageLayout() {
+    const currentTheme = useSelector(state => state.theme.theme);
+    const theme = getTheme(currentTheme);
 
     return (
         <>
-            <NavigationBar/>
-            <Outlet/>
+            <ThemeProvider theme={theme}>
+                <CssBaseline/>
+                <NavigationBar/>
+                <Outlet/>
+            </ThemeProvider>
+
         </>
     )
 }
